Consolidate logging modules
