+++ autogpt/agents/agent.py
@@ -13,13 +13,13 @@ from autogpt.llm.base import ChatModelResponse
 from autogpt.llm.chat import chat_with_ai
 from autogpt.llm.providers.openai import OPEN_AI_CHAT_MODELS
 from autogpt.llm.utils import count_string_tokens
-from autogpt.log_cycle.log_cycle import (
+from autogpt.logs import logger, print_assistant_thoughts, remove_ansi_escape
+from autogpt.logs.log_cycle import (
     FULL_MESSAGE_HISTORY_FILE_NAME,
     NEXT_ACTION_FILE_NAME,
     USER_INPUT_FILE_NAME,
     LogCycleHandler,
 )
-from autogpt.logs import logger, print_assistant_thoughts, remove_ansi_escape
 from autogpt.memory.vector import VectorMemory
 from autogpt.models.command_registry import CommandRegistry
 from autogpt.speech import say_text
