+++ autogpt/llm/chat.py
@@ -12,8 +12,8 @@ from autogpt.config import Config
 from autogpt.llm.api_manager import ApiManager
 from autogpt.llm.base import ChatSequence, Message
 from autogpt.llm.utils import count_message_tokens, create_chat_completion
-from autogpt.log_cycle.log_cycle import CURRENT_CONTEXT_FILE_NAME
 from autogpt.logs import logger
+from autogpt.logs.log_cycle import CURRENT_CONTEXT_FILE_NAME
 
 
 # TODO: Change debug from hardcode to argument
